# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alba Digital API",
        "description": "An API of ecommerce",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/order/unpaid": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Unpaid transaction",
                "description": "view a unpaid invoice",
                "operationId": "orderShow",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "total": {
                                                    "type": "string",
                                                    "example": "70000"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "unpaid",
                                                        "paid"
                                                    ],
                                                    "example": "unpaid"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "price": {
                                                                "type": "string"
                                                            },
                                                            "updated_at": {
                                                                "type": "string"
                                                            },
                                                            "created_at": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "example": [
                                                        {
                                                            "id": 1,
                                                            "name": "product 1",
                                                            "price": "10000",
                                                            "updated_at": "2023-08-28 06:06:17",
                                                            "created_at": "2023-08-28 06:06:17"
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Checkout payment",
                "description": "Checkout payment of a unpaid invoice",
                "operationId": "orderCheckout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "total": {
                                                    "type": "string",
                                                    "example": "30000"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "unpaid",
                                                        "paid"
                                                    ],
                                                    "example": "paid"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "price": {
                                                                "type": "string"
                                                            },
                                                            "updated_at": {
                                                                "type": "string"
                                                            },
                                                            "created_at": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "example": [
                                                        {
                                                            "id": 1,
                                                            "name": "product 1",
                                                            "price": "10000",
                                                            "updated_at": "2023-08-28 06:06:17",
                                                            "created_at": "2023-08-28 06:06:17"
                                                        },
                                                        {
                                                            "id": 2,
                                                            "name": "product 2",
                                                            "price": "20000",
                                                            "updated_at": "2023-08-28 06:06:17",
                                                            "created_at": "2023-08-28 06:06:17"
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}